ibot/requirements.txt
absl-py==0.12.0
addict==2.4.0
albumentations==1.1.0
anyio==3.5.0
appdirs==1.4.4
argon2-cffi==20.1.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
attrs==20.3.0
auditwheel==5.1.2
Babel==2.9.1
backcall==0.2.0
backports.entry-points-selectable==1.1.1
bayesian-optimization==1.2.0
bitstring==3.1.9
bleach==3.3.0
blis==0.4.1
boto3==1.18.21
botocore==1.21.41
Bottleneck==1.3.2
brotlipy==0.7.0
cachetools==4.2.4
carla==0.9.13
catalogue==1.0.0
certifi==2022.9.24
cffi==1.14.5
charset-normalizer==2.0.4
cityscapesScripts==2.2.0
click==8.0.3
cloudpickle==2.0.0
coloredlogs==14.0
cryptography==36.0.0
cycler==0.10.0
cymem==2.0.5
Cython==0.29.22
dataclasses==0.6
debugpy==1.5.1
decorator==4.4.2
defusedxml==0.7.1
deprecation==2.1.0
dill==0.3.3
distlib==0.3.4
distro==1.6.0
en-core-web-sm==3.1.0
entrypoints==0.3
faiss-cpu==1.6.5
filelock==3.4.0
Flask==2.0.2
flatbuffers==1.12
fonttools==4.25.0
future==0.18.2
gast==0.3.3
gensim==4.0.1
google-auth==1.28.0
google-auth-oauthlib==0.4.3
google-pasta==0.2.0
GPUtil==1.4.0
graphviz==0.5.1
grpcio==1.32.0
gym==0.21.0
h5py==2.10.0
huggingface-hub==0.10.1
humanfriendly==9.1
idna==3.3
imageio==2.12.0
imbalanced-learn==0.8.1
importlib-metadata==4.8.2
iniconfig==1.1.1
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.2
jmespath==0.10.0
joblib==1.0.1
json5==0.9.6
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.1
jupyter-packaging==0.11.1
jupyter-server==1.13.5
jupyterlab==3.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
jupyterlab-widgets==1.0.0
jupyterthemes==0.20.0
kaggle==1.5.12
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
lesscpy==0.15.0
libclang==13.0.0
llvmlite==0.39.1
lxml==4.6.3
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.2
mediapipe==0.8.3.1
mistune==0.8.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mpmath==1.2.1
multidict==5.2.0
munkres==1.1.4
murmurhash==1.0.5
nb-conda==2.2.1
nb-conda-kernels==2.3.1
nbclassic==0.3.5
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
networkx==2.6.3
neupy==0.6.5
neuralgym==0.0.1
nltk==3.5
notebook==6.2.0
numba==0.56.4
numexpr==2.7.3
numpy==1.22.2
oauthlib==3.2.0
olefile==0.46
open3d==0.14.1
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
pandarallel==1.5.2
pandas==1.2.2
pandas-datareader==0.10.0
pandocfilters==1.4.3
parso==0.8.1
pathy==0.4.0
pbr==5.5.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
pip==22.3.1
plac==1.1.3
platformdirs==2.4.0
pluggy==0.13.1
ply==3.11
preshed==3.0.5
progressbar2==3.34.3
prometheus-client==0.9.0
prompt-toolkit==3.0.16
protobuf==3.15.6
psycopg2-binary==2.8.6
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycryptodome==3.11.0
pydantic==1.5
pydot==1.4.2
pyelftools==0.28
pygame==2.1.2
PyGLM==2.5.7
Pygments==2.10.0
pyjnius==1.3.0
PyOpenGL==3.1.5
pyOpenSSL==21.0.0
pyparsing==2.4.7
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyquaternion==0.9.9
pyrsistent==0.17.3
pyserini==0.10.1.0
PySide2==5.15.2.1
PySocks==1.7.1
pytest==6.2.2
python-dateutil==2.8.2
python-slugify==5.0.2
python-utils==2.5.6
pytz==2021.3
PyWavelets==1.2.0
PyYAML==6.0
pyzmq==22.0.3
qtconsole==5.0.2
QtPy==1.9.0
qudida==0.0.4
rank-bm25==0.2.1
regex==2020.11.13
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.8
s3transfer==0.5.0
sacremoses==0.0.43
scikit-image==0.19.1
scikit-learn==0.24.1
scipy==1.6.1
seaborn==0.11.2
Send2Trash==1.5.0
sentencepiece==0.1.94
setuptools==47.3.1
Shapely==1.8.0
shiboken2==5.15.2.1
sip==4.19.13
six==1.15.0
sklearn==0.0
smart-open==3.0.0
sniffio==1.2.0
spacy==2.2.4
spacy-legacy==3.0.8
srsly==1.0.5
sympy==1.8
tableprint==0.7.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-docs==0.0.0.dev0
tensorflow-estimator==2.4.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
terminado==0.9.2
testpath==0.4.4
testresources==2.0.1
text-unidecode==1.3
tf-estimator-nightly==2.8.0.dev2021122109
Theano==1.0.0
thinc==7.4.0
threadpoolctl==2.1.0
tifffile==2021.11.2
timm==0.6.11
tld==0.12.5
tokenizers==0.13.1
toml==0.10.2
tomlkit==0.9.1
torch==1.12.1
torch-tb-profiler==0.3.1
torchaudio==0.12.1
torchtext==0.13.1
torchvision==0.13.1
tornado==6.1
tqdm==4.45.0
traitlets==5.0.5
transformers==4.23.1
typer==0.3.2
typing==3.7.4.3
typing-extensions==3.10.0.2
urllib3==1.26.7
virtualenv==20.10.0
wasabi==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
Werkzeug==1.0.1
wheel==0.37.1
widgetsnbextension==3.5.1
wordcloud==1.8.1
wrapt==1.12.1
xlrd==2.0.1
yapf==0.32.0
zipp==3.6.0
